Move several files to designated dirs
polymorphic_wrapper.hpp and visitor_apply_for_all.hpp are utility files,
and belong to the utils, rather than interfaces folder

Hashable, Signable, Primitive and ModelPrimitive were located in root of interfaces.
To keep things clean it is better to move all of them to separate directory
