+++ shared_model/interfaces/base/signable.hpp
@@ -21,8 +21,8 @@
 #include <boost/functional/hash.hpp>
 #include <unordered_set>
 #include "interfaces/common_objects/signature.hpp"
-#include "interfaces/hashable.hpp"
-#include "interfaces/polymorphic_wrapper.hpp"
+#include "interfaces/base/hashable.hpp"
+#include "utils/polymorphic_wrapper.hpp"
 #include "utils/string_builder.hpp"
 
 namespace shared_model {
