+++ shared_model/interfaces/commands/append_role.hpp
@@ -18,8 +18,8 @@
 #ifndef IROHA_SHARED_MODEL_APPEND_ROLE_HPP
 #define IROHA_SHARED_MODEL_APPEND_ROLE_HPP
 
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/primitive.hpp"
 #include "model/commands/append_role.hpp"
 
 namespace shared_model {
