+++ shared_model/interfaces/commands/set_quorum.hpp
@@ -18,8 +18,8 @@
 #ifndef IROHA_SHARED_MODEL_SET_QUORUM_HPP
 #define IROHA_SHARED_MODEL_SET_QUORUM_HPP
 
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/primitive.hpp"
 #include "model/commands/set_quorum.hpp"
 
 namespace shared_model {
