+++ shared_model/interfaces/common_objects/account_asset.hpp
@@ -19,9 +19,9 @@
 #define IROHA_SHARED_MODEL_ACCOUNT_ASSET_HPP
 
 #include <new>
+#include "interfaces/base/hashable.hpp"
 #include "interfaces/common_objects/amount.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/hashable.hpp"
 #include "model/account_asset.hpp"
 #include "utils/string_builder.hpp"
 
@@ -81,7 +81,8 @@ namespace shared_model {
         oldModel->account_id = accountId();
         oldModel->asset_id = assetId();
         using OldBalanceType = decltype(oldModel->balance);
-        /// Use shared_ptr and placement-new to copy new model field to oldModel's field and
+        /// Use shared_ptr and placement-new to copy new model field to
+        /// oldModel's field and
         /// to return raw pointer
         auto p = std::shared_ptr<OldBalanceType>(balance().makeOldModel());
         new (&oldModel->balance) OldBalanceType(*p);
