+++ shared_model/interfaces/common_objects/types.hpp
@@ -23,7 +23,7 @@
 #include <vector>
 #include "common/types.hpp"
 #include "cryptography/public_key.hpp"
-#include "interfaces/polymorphic_wrapper.hpp"
+#include "utils/polymorphic_wrapper.hpp"
 
 namespace shared_model {
