+++ shared_model/interfaces/queries/get_account_assets.hpp
@@ -18,8 +18,8 @@
 #ifndef IROHA_SHARED_MODEL_GET_ACCOUNT_ASSETS_HPP
 #define IROHA_SHARED_MODEL_GET_ACCOUNT_ASSETS_HPP
 
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/primitive.hpp"
 #include "model/queries/get_account_assets.hpp"
 
 namespace shared_model {
