+++ shared_model/interfaces/queries/get_account_transactions.hpp
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/primitive.hpp"
 #include "model/queries/get_transactions.hpp"
 
 #ifndef IROHA_SHARED_MODEL_GET_ACCOUNT_TRANSACTIONS_HPP
