+++ shared_model/interfaces/queries/get_asset_info.hpp
@@ -18,8 +18,8 @@
 #ifndef IROHA_SHARED_MODEL_GET_ASSET_INFO_HPP
 #define IROHA_SHARED_MODEL_GET_ASSET_INFO_HPP
 
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/primitive.hpp"
 #include "model/queries/get_asset_info.hpp"
 
 namespace shared_model {
