+++ shared_model/interfaces/transaction_responses/stateful_valid_tx_response.hpp
@@ -18,7 +18,7 @@
 #ifndef IROHA_STATEFUL_VALID_TX_RESPONSE_HPP
 #define IROHA_STATEFUL_VALID_TX_RESPONSE_HPP
 
-#include "interfaces/primitive.hpp"
+#include "interfaces/base/primitive.hpp"
 #include "model/transaction_response.hpp"
 #include "utils/string_builder.hpp"
