+++ shared_model/interfaces/transaction_responses/tx_response.hpp
@@ -19,8 +19,7 @@
 #define IROHA_TX_RESPONSE_HPP
 
 #include <boost/variant.hpp>
-#include "interfaces/polymorphic_wrapper.hpp"
-#include "interfaces/primitive.hpp"
+#include "interfaces/base/primitive.hpp"
 #include "interfaces/transaction.hpp"
 #include "interfaces/transaction_responses/committed_tx_response.hpp"
 #include "interfaces/transaction_responses/stateful_failed_tx_response.hpp"
@@ -28,8 +27,9 @@
 #include "interfaces/transaction_responses/stateless_failed_tx_response.hpp"
 #include "interfaces/transaction_responses/stateless_valid_tx_response.hpp"
 #include "interfaces/transaction_responses/unknown_tx_response.hpp"
-#include "interfaces/visitor_apply_for_all.hpp"
 #include "model/transaction_response.hpp"
+#include "utils/polymorphic_wrapper.hpp"
+#include "utils/visitor_apply_for_all.hpp"
 
 namespace shared_model {
   namespace interface {
