+++ shared_model/utils/polymorphic_wrapper.hpp
