+++ shared_model/utils/visitor_apply_for_all.hpp
